+++ CHANGELOG.md
@@ -95,7 +95,7 @@
 - Cache store meta data server-side and only refetch if it changed #1533 !4103 @AntonBallmaier
 - Parallelized store page loading !4103 @AntonBallmaier
 - Replaced normalisation functions in the API with DTOs !4078 @alex.simm 
-- Avatar images are now loaded lazily #1749 !4098 @AntonBallmaier
+- Avatar images are now loaded lazily #1749 !4098 #2365 !4105 @AntonBallmaier
 - Use caching for baskets #1533 !4100 @AntonBallmaier
 
 ## Dev/Test/CI stuff